Move types to types/formatted/healthCareLocalFacility
